# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: eFoods - Food Ordering & Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List all active restaurants
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Create a restaurant (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/cancel:
    patch:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Restaurant:
      type: object
      required: [id, name, description, cuisine, rating, deliveryTime, deliveryFee, minOrder, imageUrl, isOpen, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        cuisine:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        minOrder:
          type: number
        imageUrl:
          type: string
        isOpen:
          type: boolean
        address:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RestaurantInput:
      type: object
      required: [name, description, cuisine, deliveryTime, deliveryFee, minOrder, imageUrl]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        cuisine:
          type: string
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        minOrder:
          type: number
        imageUrl:
          type: string
        address:
          type: string

    # ── MENU ────────────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, userId, restaurantId, restaurantName, items, subtotal, deliveryFee, total, status, deliveryAddress, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        restaurantId:
          type: integer
        restaurantName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, preparing, out_for_delivery, delivered, cancelled]
        deliveryAddress:
          type: string
        estimatedDelivery:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [id, menuItemId, name, price, quantity, imageUrl]
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string
    OrderInput:
      type: object
      required: [deliveryAddress, razorpayOrderId, razorpayPaymentId, razorpaySignature]
      properties:
        deliveryAddress:
          type: string
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string

    # ── DASHBOARD ───────────────────────────────────────────────────────────
...